requirements.txt
asyncio==3.4.3
discord-py-slash-command==4.2.1
discord.py==2.0.1
mcstatus==10.0.1
MinePI==0.5.0
numpy==1.24.1
Pillow==9.5.0
psutil==5.9.4
pytz==2022.7.1
pyzbar==0.1.9
requests==2.27.1